# Conflicts:
#	target/m2e-wtp/web-resources/META-INF/maven/Acme.com/Acme-Handy-Worker/pom.properties

+++ src/main/java/services/MessageService.java
@@ -93,7 +93,7 @@ public class MessageService {
 				final List<Message> m1 = (List<Message>) spamBoxRx.getMessages();
 				m1.add(result);
 				spamBoxRx.setMessages(m1);
-				this.boxService.save(spamBoxRx);
+				this.boxService.saveRecipient(spamBoxRx, a);
 			}
 		else
 			for (int i = 0; i < recipients.size(); i++) {
@@ -103,7 +103,7 @@ public class MessageService {
 				final List<Message> m1 = (List<Message>) inBoxRx.getMessages();
 				m1.add(result);
 				inBoxRx.setMessages(m1);
-				this.boxService.save(inBoxRx);
+				this.boxService.saveRecipient(inBoxRx, a);
 			}
 		//Guardar mensaje en BD
 		return result;
@@ -112,7 +112,14 @@ public class MessageService {
 	public void delete(final Message message) {
 		Assert.notNull(message);
 		Assert.isTrue(message.getId() != 0);
-		this.messageRepository.delete(message);
+		final List<Box> boxes = new ArrayList<Box>();
+		boxes.addAll(this.boxService.findAll());
+		boolean msgInAnyBox = false;
+		for (int i = 0; i < boxes.size(); i++)
+			if (boxes.get(i).getMessages().contains(message))
+				msgInAnyBox = true;
+		if (msgInAnyBox)
+			this.messageRepository.delete(message);
 	}
 
 	//Other business methods
